refactor: split integration tests into smoke, anonymous, local, and helpers files
